pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "analyzer>=0.1.1",
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "kamene>=0.32",
    "numpy>=2.2.4",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.39",
    "stereoids>=0.2.9",
    "flask-wtf>=1.2.2",
    "twilio>=9.5.1",
    "flask-bcrypt>=1.0.1",
    "flask-mail>=0.10.0",
    "wtforms>=3.2.1",
    "werkzeug>=3.1.3",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
    "openai>=1.70.0",
    "cairosvg>=2.7.1",
]
